server/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jwcrypto==1.5.6
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2

# LLM and AI functionality
openai==1.58.1
anthropic==0.40.0
langchain==0.3.12
langchain-openai==0.2.14
langchain-anthropic==0.2.7
langchain-community==0.3.12

# Document processing
PyPDF2==3.0.1
python-docx==1.1.2
python-multipart==0.0.17
aiofiles==24.1.0

# Vector storage and embeddings
chromadb==0.5.23
sentence-transformers==3.3.1
numpy==2.2.1

# Text processing
tiktoken==0.8.0
beautifulsoup4==4.12.3
lxml==5.3.0
